+++ flake.nix
@@ -69,6 +69,7 @@
           {
             services.clawpi.audio.enable = true;
             services.clawpi.audio.groq.enable = true;
+            services.clawpi.elevenlabs.enable = true;
             services.clawpi.telegram = {
               enable = true;
 
@@ -101,6 +102,7 @@
             services.clawpi.debug = true;
             services.clawpi.audio.enable = true;
             services.clawpi.audio.groq.enable = true;
+            services.clawpi.elevenlabs.enable = true;
             services.clawpi.telegram = {
               enable = true;